SWARM-1153 - Be more careful about -Dswarm.logging.** property setting.
Motivation
----------

There's a lot of properties that are not setting category logging
levels, and we should be better about not breaking.

Modifications
-------------

Not only checking for simple logger name, but also seeing if the RHS of
the property represents a log-level.  If not, it's probably not a logger
category.

Plus, tests.

Result
------

Less breakage.