SWARM-1155: Datasource test should not rely on JDK7
Motivation
----------
When using EAP the JDK8 classes are used instead, causing this test to fail. Need to make the test pass with JDK7 or JDK8 classes.

Modifications
-------------
Modified the test to use a regex comparison that works with either JDK7 or JDK8.

Result
------
No impact to product, only a test.