SWARM-1163: Provide interface for version, etc (#405)
Motivation
----------
Currently the WF Swarm version is only available as a system property. We want to make it, and other details, available an interface.

Modifications
-------------
Created SwarmInfo that is filtered during build to add the version and groupId.

Result
------
No impact to product functionality, but version is available through SwarmInfo.