SWARM-1166: JUnit Test categories
Motivation
----------
We need to be able to exclude tests from community or product builds due to differing functionality.

It's now possible to add `@Category(CommunityOnly.class)` or `@Category(ProductOnly.class)` onto a Test or Test method and have it excluded from the other build type.

Modifications
-------------
Added `build-resources` module to contain the interfaces defining the categories, modified the surefire-plugin configuration to set `excludedGroups` based on property that is set by community and product builds.

Result
------
No impact on product, purely for tests.