SWARM-1036 - Management interface is terribly open by default.
Motivation
----------

Since the management fraction is required by the undertow fraction
if you're attempting to use HTTPS, it should *not* bind to the public
interface by default.

Modifications
-------------

When the management fraction is included, we now behave more like
WildFly proper by creating a `management` interface that, by default,
binds to 127.0.0.1 (instead of 0.0.0.0/all) and is used by the
management socket-bindings for HTTP and HTTPS interfaces (9990 etc).

This can be adjusted using swarm.management.bind.address, if the
desire is to use a different address.

Additionally, the management interface can be ignored, and reset
to use the public interface (possibly ill-adviced) by the configuration
item of swarm.management.bind.interface=public instead.

Result
------

Safer and sturdier.