SWARM-1087: Adjust to handle product config-api (#411)
Motivation
----------
Our productized wildfly-config-api doesn't include a means to generate a self signed certificate so we need to handle cases where that method isn't present.

Modifications
-------------
Throw an error when generate self signed cert requested and it's not available. Switch to calling the method through reflection to handle when it isn't there.

Result
------
No direct impact.