SWARM-1220: Support propertiesFile loading from classpath/URL.
Motivation
----------
In certain maven setups, the properties file resides in a different module than the swarm application/plugin.
Those setups need to make extra effort so that properties file can be used as a regular file path by the plugin(s).

Modifications
-------------
Extend PropertiesUtil.loadProperties(String) to load from (in that order):
- a regular file path (unchanged)
- the classpath via thread context classloader
- a URL

Result
------
A propertiesFile can now be alternatively be loaded:
- from a plugin dependency -> classpath resource
- from a jar:file-URL
- a "remote" URL/server

The same applies to environmentFile as it is processed in the same way as propertiesFile.