SWARM-1170 - If it's a war, we should auto-detect undertow.
Motivation
----------

Sometimes (I guess?) it's possible to have a .war that contains
nothing undertow/servlet/jaxrs-like within it, yet still function.

Modifications
-------------

Since FractionDetectors work on the *content* of the files, add
a special-case (for now) for .war == undertow.

Result
------

A .war without anything webby is known to be a war and detects
and requests undertow.