SWARM-1231 - When management and webservices is used, Ambgiuous CDI.
Motivation
----------

When :management fraction is included, we setup multiple interfaces.
When WebServices is used, it injects an interface, which is now ambiguous.

Modifications
-------------

Mix in some @Named annotations for both the creation and injection
of interfaces to avoid ambiguity.

Result
------

Should be able to inject Interface.PUBLIC or Interface.MANAGEMENT (if
present) as needed, unambiguously.