SWARM-1249: MavenPluginTest should pass through custom settings.xml setting
Motivation
----------
When the `MavenPluginTest` invokes Maven (via the `Verifier`), it should
pass through the location of `settings.xml`, if there's one configured.

This is because the testing Maven project can depend on artifacts that
are only available in a custom Maven repository, which is configured
in the `settings.xml` (e.g. a productized Swarm build).

Modifications
-------------
Detect if `settings.xml` location is configured (by checking a system
property that Maven sets) and if so, pass that to the `Verifier`
for use when invoking Maven.

Result
------
It is now possible for the testing project to require artifacts
from a custom repository configured in custom `settings.xml`,