SWARM-1217: Do not add generated Application if servlet mapping exists.
Motivation
----------
According to the JAX-RS spec if no Application subclass is present and there is web.xml that specifies a
servlet mapping for a servlet with name javax.ws.rs.core.Application, the servlet is added automatically.
In this case, Swarm must not add a generated Application class.

Modifications
-------------
JAXRSArchiveImpl now attempts to analyze the existing web.xml and does not add/removes the generated Application class if a mapping is present.

Result
------
jboss-eap-quickstarts/helloworld-html5 should pass.