SWARM-1298: avoid double logging when using swarm.logging.category system properties
Motivation
----------
If I use `-Dswarm.logging.category.name=DEBUG` to enable
debug logging, I get each of the debug log messages twice.
This appears e.g. in the `ArqLoggingLevelsTest` test.

The ultimate reason is that `LoggingFraction.applyDefaults(Level)`,
which adds the logger to the WildFly `logging` subsystem,
configures a handler for the logger. The thing is that by default,
log messages are already sent to the parent's logger handlers,
so in combination, the log message gets sent to 2 handlers.

Modifications
-------------
The newly created logger no longer has a handler set up.
This is how loggers are typically added with WildFly.

Another possibility would be to set the `use-parent-handlers`
management attribute to `false`.

Result
------
No more double logging.