SWARM-1223: Add test for SWARM-1223.
Motivation
----------
It's advisable having a test for new features to ensure this is not regressed in future.

Modifications
-------------
New JUnit test had been added, checking with FractionUsageAnalyzer (same class used for fractions in the end) that jpa fraction exists when adding a persistence.xml file

Result
------
After the test, jpda fraction detection should be ok and test should pass.