SWARM-1322: Swarm's Arquillian container doesn't enable the @EJB enricher (#492)
Motivation
----------
Arquillian ships with a bunch of test instance "enrichers"
for injecting objects to annotated fields. Among them are
enrichers for Java EE annotations `@Inject`, `@Resource`
and `@EJB`. These enrichers are enabled only if the respective
annotation (or some other identifying class) is "present
on the classpath". Which, for in-container tests, means
_visible from the classloader that loaded Arquillian_.
In Swarm's case, that's the module classloader for module
`org.jboss.arquillian`. This module is defined by Swarm itself.

This module currently exports `javax.annotation.api`, which
contains `@Resource`, and `javax.enterprise.api`, which
contains `@Inject` (and the `BeanManager`). For the `@EJB`
enricher to work, an export of `javax.ejb.api` is needed.

Modifications
-------------
Added an exported dependency on the `javax.ejb.api` module
to the `org.jboss.arquillian` module.

Result
------
`@EJB` enricher can actually work with Swarm Arquillian.