ENTSWM-31: added switches to download sources and poms
Motivation
----------
Prod needs to build a repository that contains sources and poms for the artifacts used by a user app.

Modifications
-------------
`ExtraArtifactsHandler` has been added and invoked in `MavenArtifactResolvingHelper#resolveAll`.
Based on system properties, pom, javadoc or source artifacts are added to the list of `DependencyNode`s to resolve.

Result
------
When `-Dswarm.download.poms`, `-Dswarm.download.javadocs` or `-Dswarm.download.sources`, is set during a build of a user app, the pom files, javadocs or sources (appropriately) are downloaded for every artifact downloaded by the swarm plugin.