SWARM-1222: Fix security web xml detector and add security fraction detection test.
Motivation
----------
The initial motivation of this commit was adding test for security detection. After doing it, it was noticeable that security web.xml detector wasn't working, so another fix was needed.

Modifications
-------------
After fixing security detector, other bug arised, because shrinkwrap utility class for parsing web.xml (Descriptors) was closing the stream, which caused error when a second webxml fraction detector was executed (for example, JSFServletDetector and SecurityWebXmlDetector). In order to fix that, it was need to use inputstream that needed 2 requirements:
 1-Not closing stream just in case any used tool was doing it without asking for it (like Descriptors class).
 2-Enabling support for marking inputStream, as we need to "rewind" stream to the initial point so next webxml fraction detectors is chain could do their job.

Although the new inputstream with previous requirements (UncloseableBufferedInputStream) was added at the generic scanning method (we need the same inputstream so we can reset to the beginning of the stream), it shouldn't affect other detectors that doesn't use mark and reset. These features are used in SecurityWebXmlDetector, though, where we mark the inputstream each time, so we can reset the stream if it reached its end (stream.available ==0).

Result
------
New test is added, checking that security feature won't be regressed if any change is done. Additionally, now all security detection is being made correctly. Last improvement, we now have a mechanism just in case we suffer for the same issue in other fraction detector.