SWARM-1364: Make versions used only in module.xml's applicable for alignment
Motivation
----------
Dependency versions that are used only in module.xml's are not aligned by the PME.
The point of this change is to make PME align them.

Modifications
-------------
The version properties that were defined in the top pom, have been given an entry in the dependency management section.
The versions from specific modules have now `provided` dependencies.

Result
------
All the versions but jboss-msc get aligned. jboss-msc requires further investigation