SWARM-1222: Fix security web xml detector, add security fraction detection tests and cleanup API. (#514)
Motivation
----------
The initial motivation of this commit was adding test for security detection. After doing it, it was noticeable that security web.xml detector wasn't working, so another fix was needed.

Modifications
-------------
After fixing security detector, other bug arised, because shrinkwrap utility class for parsing web.xml (Descriptors) was closing the stream, which caused error when a second webxml fraction detector was executed (for example, JSFServletDetector and SecurityWebXmlDetector). In order to fix that, some steps had been taken:

1-Unify under FileSource (renamed to PathSource to make more consistent to their new inherit classes), now one more method had been added to make easier obtaining InputStream. PathSource has 2 inherit classes: FilePathSource (equivalent to previous FileSource) and ZipPathSource.
2-Thanks to previous unification, delegate InputStream obtaining to Scanner or Detector classes.
3-In order to make classes just doing a single purpose ("Separation of concerns"), code for traversing files, directories and zip entries from Scanner had been moved to FractionUsageAnalyzer. This made possible removing many code from Scanner class. So now FractionUsageAnalyzer just does File/zip traversing, executing other methods, then delegate simple source scan to Scanner.
4-Now that FileSource (named PathSource) has everything we want, unify all methods to just accept PathSource instead of the previous ones (InputStream, String ,etc).

Result
------
Now SecuredWebXMLDetector is working, and API in detectors and Scanners seems more consistent and unified.