SWARM-1386 - Handle maven mirror better.
Motivation
----------
Because multiple repositories that were mirrored by a single
mirror resulted in trampled maven-metadata.xml, there are times
that some artifacts could not be resolved.

Modifications
-------------
Ensure that if a repository is added to our tool and is wrapped
in a mirror, that the resulting wrapped repository contains the
ID of the inner repository, and not the mirror's ID.

Result
------
JSF example should now pass.