SWARM-1411 - Added Jaeger Fraction
Motivation
----------
This Fraction brings the Jaeger and OpenTracing dependencies and gives
a Swarm-way to configure the tracer.

The main advantage of using this fraction vs. the OpenTracing fraction
plus the Jaeger dependencies is that this brings the correct versions of
all the components required to have a fully working OpenTracing solution.

Modifications
-------------
* Added Jaeger fraction
* Added a property to the OpenTracing fraction to skip automatic resolution
of tracers

Result
------
Applications consuming this fraction now get automatic and effortless support
for a complete OpenTracing solution, including automatic capture of request
data, thanks to the TracingFilter from the OpenTracing fraction.