SWARM-1426 - Allow specification of deployment export directory.
Motivation
----------
Sometimes when using swarm.export.deployment, $PWD isn't the
best place to export the deployment.

Modifications
-------------
swarm.export.deployment, if set to "true" (or without a value),
continues to export in $PWD.  If set to some other value,
that value will be treated as a path to a directory where
the deployment should be exported.  If that directory does
not exist, it will be created.

Result
------
-Dswarm.export.deployment=./some/dir works as you might hope.