SWARM-1434: Move dependencyManagement to unsupported profile
Motivation
----------
If a dependency defined in dependencyManagement section is a pom that is imported, it needs to be in the appropriate profile otherwise it brings in dependencies into the local repo that are not actually needed.

Modifications
-------------
Moved vertx-dependencies into dependencyManagement section of unsupported profile.

Also reorganized some dependencies in default profile to be with "like minded" dependencies, to make it easier to grasp the needed dependencies for pieces.

Result
------
No impact on behavior, only modifies build.