SWARM-1447: default datasource not set in the EE subsystem
Motivation
----------
When a default datasource is created (see
`DatasourceAndDriverCustomizer.customizeDefaultDatasource`),
it doesn't get set as the default datasource in the `ee` subsystem.
This is the primary place where WildFly goes to for the default
datasource.

Since databases are most often accessed via JPA, this issue wasn't
visible much, because the `jpa` fraction takes that default datasource
from the `datasources` fraction and sets it as a default for the `jpa`
subsystem. This means that JPA works with the default datasource just
fine, but other parts of the application server don't (e.g. injecting
`@Resource DataSource` into a servlet).

Modifications
-------------
The default datasource is set in the `ee` subsystem now.
It is no longer set in the `jpa` subsystem, which is now forced
to consult the `ee` subsystem.

Result
------
Default datasource, if exists, is properly available to the entirety
of the application server. This improves Java EE 7 compatibility.