Fix ReflectionUtil.getMethodsWithAnnotation() (#577)
Fixed the function to also list annotated methods of subclasses of the queried class.

The bug appeared when creating a base class for integration-tests that handles the creation of Swarm in a central point in code. To enable the use of DRY and SRP in tests, a superclass should handle container/deployment - setup.