SWARM-1476: topology-webapp formats a service URL with an IPv6 address incorrectly
Motivation
----------
A URL with an IPv6 address should look like
`http://[my::ad:dr:es:s]:8080/`, but the SSE stream maintained
by the `topology-webapp` fraction doesn't add the brackets.
The URL looks like `http://my::ad:dr:es:s:8080/`, which is wrong
and doesn't work in browsers.

Modifications
-------------
Attempt to detect if the service address is an IPv6 address,
and if so, add the brackets.

Result
------
`topology-webapp` works in an IPv6 environment.