SWARM-1472: Arquillian fraction has test scope in all boms.
Motivation
----------
Arquillian fraction should have "test" scope in all boms so that its
dependencies are not added to the resulting artifact during compilation.

Modifications
-------------
Added <swarm.fraction.scope>test</swarm.fraction.scope> to the
arquillian fraction. The property will be currently ignored (arquillian
fraction is handled in a special way) but it's good to be consistent.

Result
------
Nothing changes.