SWARM-1526: topology.js should have a mechanism to wait for initial topology
Motivation
----------
When `topology()` is called, it subscribes to the SSE stream,
which immediately sends back the current topology. There
should be a mechanism to wait for that, otherwise users are
susceptible to race conditions.

Modifications
-------------
The `topology()` function no longer returns the `Topology` object.
Instead, it returns a promise that is resolved with the `Topology`
object once the first event from the stream arrives. That way,
the race condition can never occur, the user can only get hold
of the `Topology` object after the initial topology is obtained.

This is obviously a breaking change. Migration path is pretty simple,
though; instead of `var topo = topology(); ... use topo ...`, users
should use `topology().then(function(topo) { ... use topo ... });`.

Result
------
Users are no longer susceptible to race conditions with topology.js.
Of course, they have to migrate to the new solution.