SWARM-1033 -- Support usage.txt
Motivation
----------
Some -swarm.jar apps might enjoy being able to display some
usage information after booting.

Modifications
-------------

If a user's application has one of:

* `usage.txt`
* `META-INF/usage.txt`
* `WEB-INF/usage.txt`

it will be displayed immediately prior to the message that
states "WildFly Swarm is Ready".

Variables within this file, in the form of ${foo} will be interpolated
if the associated configurable is available.  This is useful for
pointing out which port or URL might be available, such as by using
${swarm.http.port} within the file.

If the configurable is un-set, then the variable will *not* be
interpolated, and you will see a literal `${whatever}` in the
output.  `usage.txt` is *not* intended to be a fancy scripting
language. The variables *are* exposed from defaults after all
configurables have been processed from the fraction configurations,
so they do not need to be explicitly set.

Result
------
Ability to display customized usage information at boot.