SWARM-1536: topology-openshift uses wrong namespace when pod runs in a different namespace than the image was built in
Motivation
----------
`org.wildfly.swarm.topology.openshift.runtime.NamespaceService`
contains a simple algorithm to select which namespace to target
when reading the topology. This checks the `KUBERNETES_NAMESPACE`
env var, then `OPENSHIFT_BUILD_NAMESPACE` and then possibly
the list of all projects iff it contains a single entry.

However, the `KUBERNETES_NAMESPACE` env var isn't a feature
of Kubernetes or OpenShift. It only gets added to the pod template
by the Fabric8 Maven plugin, using the Kubernetes Downward API.
If the deployment method used doesn't involve Fabric8 Maven plugin,
this env var will be missing. Then, the `OPENSHIFT_BUILD_NAMESPACE`
env var contains the name of the namespace _where the image was
built_, which doesn't have to be the name of the namespace in which
the pod _is currently running_ (though it's probably quite common).

It seems reading `/var/run/secrets/kubernetes.io/serviceaccount/namespace`
is the best option, see also
https://kubernetes.io/docs/tasks/administer-cluster/access-cluster-api/#accessing-the-api-from-a-pod

Modifications
-------------
Changed the `NamespaceService` to try reading
`/var/run/secrets/kubernetes.io/serviceaccount/namespace`
if the `KUBERNETES_NAMESPACE` env var is missing.

Result
------
If the pod is running in a different namespace than the image
was built in, and Fabric8 Maven plugin wasn't involved, the namespace
is still selected correctly.