SWARM-1541: Unable to recognize -fae on command line
Motivation
----------
If -fae is present on a Maven command line, FileSystemLayout is unable to determine whether we're dealing with Maven or Gradle.

Modifications
-------------
Altered the -f handling to ensure that -f is only "found" when a .xml file is present as part of the command.

Result
------
Fix error when -fae is present on command line.