SWARM-1535 - Handle deployment exceptions via Arquilian. (#632)
Motivation
----------
Apparently a @Deployment can @ShouldThrowException, and we
gleefully ignored that fact, and never transmitted deployment
exceptions across to the client side.

Modifications
-------------
Capture deployment exceptions (via Service<T> start failures)
to the Daemon and thence to the client.  Have the client check
on the deployment during the deploy() phase once the Uberjar
container has decided that deployment is complete in one form
or another.

Result
------
Exceptions from @Deployment are serialized back to the test.