SWARM-1549: enable PermitAll semantics.
Motivation
----------
The current SecurityConstraint/WebXmlAsset does not carry enough information
to allow one to configure a @PermitAll semantics.

Modifications
-------------
A permitAll flag has been added to SecurityConstraint and WebXmlAsset openStream
will not call getOrCreateAuthConstraint() for a SecurityConstraint that has
permitAll true as this is what indicates uncontrained access to the given
web resources.

Result
------
This change allows one to configure unrestricted access for a collection of
url patterns, something that was not possible previously.

Signed-off-by: Scott Stark <starksm64@gmail.com>