SWARM-1414: Adding timestamp to hollow jar prevents caching as Docker layer
Motivation
----------
This commit removes the generation of the timestamp in "META-INF\wildfly-swarm-manifest.yaml". This is necessary to make the created JAR idempotent.

Modifications
-------------
In BuildTool.java the generation of the property "swarm.uberjar.build.timestamp" has been removed.

Result
------
This is the first step in making the created JAR idempotent. The timestamp does not seem to be consumed by Wildfly-Swarm itself