SWARM-1590 - Support project-*.yaml within hollow-jars.
Motivation
----------
If a project-defaults.yaml is used with a deployment and a hollow
jar, it was not recognized.

Modifications
-------------
If running as a hollow jar, support locating project-*.yaml configuration
within the _first_ deployment passed.

Result
------
The first (typically only) deployment may contain YAML to configure the
runtime of a hollow jar.