SWARM-1641: Fix MP metrics fraction.
Motivation
----------
We need to fix MP Metrics fraction to pass the TCK.

Modifications
-------------
This commit contains a lot of changes but the main change in the
integration code is that we don't add "metrics-api" directly to the
deployment but instead add a module dependency.

Result
------
MP metrics TCK should pass. This commit also fixes SWARM-1639 and
partially SWARM-1640.