[SWARM-1710] Fix test-jar
Found a problem when using wildfly swarm with arquilian and I want to use one or more classes of a dependecy's test source code. This scenario I can just import the dependency in maven specifying the type as test-jar. This will make the test source code of that depdency available to me. The problem fixed was that it was using the type to get the specific file extension but in this scenario is not the case