SWARM-1784: fix ignore datasource jndi name (#765)
Motivation
----------
The purpose of this commit is to make swarm less rigid about JNDI naming
conventions for JPA data sources.

Modifications
-------------
I have modified the producer method for the datasource so that it will
produce a jndi name instead of the name. That way, if the datasource
jndi is specified and is not in the form `java:jboss/datasource/$name`
it will still work.