Add support for easy ephemeral ports in tests.
Add @EphemeralPorts to your test that uses the TestRunner,
and regardless of your app config, it will ensure an ephemeral
port is used.  This is great for doing parallel builds.

Additionally, make RestAssured.baseURL point to the primary
URL in all cases, to make life a little easier when testing.