SWARM-1870: Allow hollow packaging mode in gradle plugin
Motivation
----------
The current maven plugin allow for hollow packaging mode.
For gradle users, this feature is not currently present.

Modifications
-------------
A 'hollow' gradle option has been added and used in the gradle packaging task.

Result
------
As a result of a 'gradle build', a *-hollow-swarm.jar file will be created along with a skinny war.