SWARM-1871: Add a fraction for Java MVC 1.0 with Ozark-RI
Motivation
----------
It would be nice to have a fraction for Java MVC, which allows to create
web UIs in a similar fashion to jax-rs services.

Modifications
-------------
This commit adds a new fraction "mvc" in the javaee folder and adds that
fraction to the parent pom.

Result
------
The new fraction will be available under the "web" tag.