SWARM-1895: [Swagger] Align library & module dependencies.
Motivation
----------
The Swagger fraction is pulling in an older version of the Guava dependency (18.0) while the Maven project lists the provided dependency to "20.0". This causes weird issues when using the Gradle plugin for building a "bundled" uber jar file as the dependency is never satisfied. Instead of downgrading the Guava dependency or messing around with the swarmtool / gradle plugin, it makes more sense to add an explicit dependency on the Guava module instead of keeping track / updating the dependency version for each release.

Additionally, at some point in time, if we have the ability to import POM properties of a BOM, we can reuse the library versions easily by referring to the appropriate property names. At a minimum, by aligning the property names, it will help in ensuring that we are copying over the same values from the appropriate bom.

Modifications
-------------
* Refer to the Guava & Javassist dependency modules instead of managing the dependency on our own.
* Align the Jackson library versions with those present in WF parent.
* Align the property name for library versions with those present in WF parent.

Result
------
Uber-bundled jars built using the Gradle plugin will no longer complain of missing Javassist, Guava & Jackson libraries.