SWARM-1790: MicroProfile + JPA Hollow Jar (#835)
Motivation
----------
As MicroProfile doesn't offer any kind of persistence at present, it would be good to offer a Hollow jar that combines MicroProfile and JPA for that purpose.

Modifications
-------------
Removed `ManifestComparitor` as it was causing comparison inconsistencies and was not really serving a purpose.

Add `microprofile-jpa` hollow jar.

Modify `ApplicationModuleFinder` to allow adding arbitrary libraries to the `swarm.application` module by using the `swarm.classpath` system property. This makes it possible for Hollow Jars to specify a path for JDBC Drivers such that they can be auto discovered.

For example, this new Hollow Jar can be run with:

  java -Dswarm.classpath=./h2-1.4.196.jar -jar microprofile-jpa-hollow-swarm.jar myApp.war

and the H2 Driver will be auto discovered.

Result
------
Added MicroProfile + JPA Hollow Jar to build, and added support for -Dswarm.classpath to specify jars to be added to application module classloader.