Simplify the POM structures
Motivation
----------
I was trying to test out some changes locally. In order to avoid any locally cached files, I decided to tweak the version literal and build a complete new set of files. As I would've done earlier (for other projects), I executed the command "mvn versions:set". Turns out, not all projects are managed by the same parent, so I still had modules whose version numbers hadn't changed. This means, that anyone trying to execute a similar exercise will have to rely on a find-and-replace-all capability instead of using predictable tooling.

The changes made in this commit improve the maintainability of the POM files.

In order to try out changes locally, all one needs to do is,

    $> mvn versions:set -DnewVersion=X.Y.Z
    $> mvn clean install

    Consume your newly built versions in your project.

Modifications
-------------
1. Replaced explicit version declarations for modules (in the dependencyManagement section) with ${project.version}
2. When referencing tools / binaries built from this very same project, make use of ${project.version} instead of the version literal.
3. Restructured the parent projects for a few of the modules.
4. Moved plugin configurations to the respective parent-modules.

Result
------
There is no impact to the project, unless there is some tooling (not listed within the project) that relies on the current POM design.