SWARM-1905: the MicroProfile TCKs setup requires bom-all (#843)
Motivation
----------
The setup of MicroProfile TCKs requires `bom-all`.
This is unnecessary and also potentially harmful, if the TCKs
were for some reason to depend on things that are outside of
the plain `bom`. It also makes it unnecessarily hard to run
the MicroProfile TCKs in the `swarm.product.build` profile.

Modifications
-------------
The MicroProfile TCKs setup is changed to depend on pure `bom`.

Result
------
MicroProfile TCKs will not inadvertently depend on artifacts
outside of `bom`. MicroProfile TCKs are easier to run
in the `swarm.product.build` profile.