SWARM-1906: JAX-RS Fraction does not introspect jar files inside war files when searching for the presence of ApplicationPath
Motivation
----------
I noticed that there were multiple JAX-RS applications deployed for my WAR file. As I debugged further, I realized that the issue was because the JAR file (inside the WAR file) was being viewed as a ByteArrayAsset and was being skipped over. With the recent fixes to IndexView representation, the fix was actually quite simple.

Modifications
-------------
1. During the deployment process, query the IndexView to see if there are **any** classes which contain the @ApplicationPath annotation.
2. If present, do not generate the default JAX-RS Application class.

Result
------
1. Verified a locally patched build and ensured that things are working as desired.
2. Ensured that the build has no errors.