SWARM-1471: Do not auto-generate the JAX-RS application if the application-path has not been defined in YAML.
Motivation
----------
Currently WF Swarm auto-generates a JAX-RS application class which leads to an inconsistent behavior (does not comply with the JAX-RS spec). Developers tend to get confused with the presence of multiple Application references being printed in the console. The intention of this auto-generation was to help developers easily define the JAX-RS application-path declaratively without needing to define a superfluous empty class. As discussed in the Jira issue, we should be generating the class only if the declarative parameter is defined.

Modifications
-------------
1. Add the Application subclass only if no sub-class of Application is present **and** the YAML file defines the "application-path" entry.
2. Updated the deployment definitions for various test cases that were relying on this implicit behavior.
3. Updated the howto projects to include an Application subclass.

Result
------
Verified the behavior (across many, many test cases).