SWARM-1925: Improve determination of the packages that need to be scanned (#858)
Motivation
----------
Swagger looks for the 'PACKAGES' key for determining the set of packages that need to be included as part of the scanning process. If no value has been specified, then it makes a rudimentary search to determine the base package that needs to be included.

The implications of the default behavior are:

* REST resources are scanned within WEB-INF/classes only.
* Classes added to library jar files that belong to a different packages are not included.

In most cases, this is not really problem as people quickly learn to configure the 'PACKAGES' parameter in the `swagger.swarm.conf` file. From a product perspective, it would be nice to include a more sensible default.

Modifications
-------------
* Scan for @Api & @Path annotated classes & methods.
* For each class identified, add its package to the list of candidate packages.
* Prune through the packages to keep only the top-level packages.
* Set the pruned collection on to the Swagger bean.

Result
------
* Verified that the build is going through properly.
* Generated a custom build and verified that the new behavior is working properly.