SWARM-1929: MP FT - circuit breaker never opens if there is a fallback.
Motivation
----------
There is a bug in MP FT implementation.

Modifications
-------------
Fix HystrixCommandInterceptor and DefaultCommand to support this basic
scenario. Also add basic support for CircuitBreaker.failOn() which was
so far completely ignored. We should align the functionality once the
spec issue is resolved.

Result
------
The bug should be fixed now.