SWARM-1947: GradleFileSystemLayout does not support customization of the build root.
Motivation
----------
If a developer is using Gradle as the build tool and decides to customize the "build" directory of the project, then Arquillian tests start failing with the message:

    WFSWARM0012: Cannot identify FileSystemLayout for given path:

This is because the `GradleFileSystemLayout` looks for content in the default folder locations and is not customizable. It would be nice if the value can be customized and provided as system properties during the test execution stage.

Modifications
-------------
In GradleFileSystemLayout,

* Look for two system properties: `src.dir` & `build.dir`.
* If the properties have not been defined, then default them.
* Use the resolved values for locating additional resources.
* Collapsed the logic of Gradle4FileSystemLayout in to this class.

Additionally,

* Updated the dependency adapter for the upcoming Gradle BOM support.
* Prepare the Arquillian Gradle Adapter to handle multi-module project resolution. This is needed for the Gradle tooling revamp.

Result
------
* Verified that the build is going through properly.
* Generated a custom build and verified that the new behavior is working properly.